+++ p1/PCEngine/src/main/java/es/ucm/vdm/pcengine/PCInput.java
@@ -3,26 +3,28 @@ package es.ucm.vdm.pcengine;
 import java.awt.event.KeyEvent;
 import java.awt.event.KeyListener;
 import java.awt.event.MouseEvent;
-import java.awt.event.MouseMotionListener;
 import java.awt.event.MouseListener;
+import java.awt.event.MouseMotionListener;
 import java.util.ArrayList;
 
-
+import es.ucm.vdm.engine.EventType;
 import es.ucm.vdm.engine.Input;
+import es.ucm.vdm.engine.TouchEvent;
 
-public class PCInput extends Input implements  MouseListener, KeyListener, MouseMotionListener {
+public class PCInput extends Input implements MouseListener, KeyListener, MouseMotionListener {
 
-    public PCInput(){
-        mTouchEvents_= new ArrayList<TouchEvent>();
+    public PCInput() {
+        mTouchEvents_ = new ArrayList<TouchEvent>();
     }
+
     @Override
     public void keyTyped(KeyEvent keyEvent) {
-
+        TouchEvent t = new TouchEvent(0, 0, EventType.KEY_DOWN);
     }
 
     @Override
     public void keyPressed(KeyEvent keyEvent) {
-
+        TouchEvent t = new TouchEvent(0, 0, EventType.KEY_DOWN);
     }
 
     @Override
