+++ p1/engine/src/main/java/es/ucm/vdm/engine/EventType.java
@@ -1,0 +1,10 @@
+package es.ucm.vdm.engine;
+
+public enum EventType {
+    PRESSED_DOWN,
+    PRESSED_UP,
+    MOVED,
+    CLICKED,
+    KEY_DOWN,
+    KEY_UP
+}
