--- p1/engine/src/main/java/es/ucm/vdm/engine/Graphics.java
@@ -1,6 +1,0 @@
-package es.ucm.vdm.engine;
-
-public abstract class Graphics implements IGraphics{
-
-
-}
